feat(content): enforce content rules
