+++ CONTENT_RULES.md
@@ -1,3 +1,60 @@
 # Правила ведения заметок
 
-Каждая заметка должна быть атомарной, дискретной, связанной с другими заметками и иметь YAML frontmatter с полями `title` и `tags`.
+Эти правила обязательны для владельца, контрибьюторов и ИИ-агентов, изменяющих `content/`.
+
+## Атомарность
+
+Каждая содержательная заметка отвечает за одну конкретную тему, понятие, операцию, проблему, шаблон или рецепт. Независимые темы выносятся в отдельные файлы и связываются ссылками. `index.md` является навигационным исключением.
+
+## Дискретность
+
+Заметка выражает полную и законченную самостоятельную мысль. Ссылки дополняют материал, но не заменяют отсутствующее объяснение.
+
+## Связанность
+
+Каждая заметка имеет хотя бы одну осмысленную входящую или исходящую ссылку. Новая заметка обычно добавляется в ближайший `index.md`.
+
+## YAML frontmatter
+
+Каждый Markdown-файл начинается с YAML-предисловия:
+
+```yaml
+---
+title: Диагностика SSH
+tags:
+  - ssh
+  - troubleshooting
+---
+```
+
+Обязательны непустой `title` и непустой список `tags`. Теги пишутся в lowercase, не повторяются и не содержат пустых элементов.
+
+## Каталоги и index.md
+
+Каждый обычный каталог внутри `content/` содержит `index.md`. Служебные каталоги, название которых начинается с `!`, не требуют `index.md` и не включаются в навигационные индексы.
+
+## Ссылки
+
+Используются только относительные Markdown-ссылки. Wiki-ссылки Obsidian не используются. Ссылка на заметку содержит `.md`, путь URL-кодируется и ведёт на существующий файл.
+
+```markdown
+[Диагностика SSH](Диагностика%20SSH.md)
+```
+
+## Вложения
+
+Изображения и другие вложения хранятся в `!assets` той же директории, где находится использующая их заметка.
+
+```markdown
+![Схема зависимостей](!assets/Схема%20зависимостей%20unit-%D1%84%D0%B0%D0%B9%D0%BB%D0%BE%D0%B2.png)
+```
+
+Имя вложения должно описывать его содержимое. Имена вроде `Pasted image`, `image`, `screenshot`, `file` и `Untitled` недопустимы.
+
+## Проверки
+
+`npm run content:links` проверяет YAML, теги, ссылки, граф заметок, размещение и имена вложений. CI дополнительно проверяет наличие `index.md` во всех обычных каталогах и пропускает каталоги `!…`.
+
+Атомарность, дискретность и смысловая полезность связей оцениваются автором и рецензентом вручную.
+
+`npm run sync` выполняет локальную проверку до Git-операций. Флаг `npm run sync -- --no-check` отключает только локальную проверку; проверки Pull Request и деплоя остаются обязательными.
